test(DatetimePicker): add demo test
