+++ example/app.json
@@ -147,6 +147,7 @@
     "van-tree-select": "./dist/tree-select/index",
     "van-tree-select-demo": "./dist/tree-select/demo/index",
     "van-datetime-picker": "./dist/datetime-picker/index",
+    "van-datetime-picker-demo": "./dist/datetime-picker/demo/index",
     "van-rate": "./dist/rate/index",
     "van-rate-demo": "./dist/rate/demo/index",
     "van-collapse": "./dist/collapse/index",
