--- example/pages/datetime-picker/index.wxss
@@ -1,1 +1,0 @@
-
