+++ packages/datetime-picker/demo/index.json
@@ -1,0 +1,8 @@
+{
+  "component": true,
+  "usingComponents": {
+    "van-toast": "../../toast/index",
+    "van-datetime-picker": "../../datetime-picker/index",
+    "demo-block": "../../../example/components/demo-block/index"
+  }
+}
